Adding image classification

+++ vision/image_classification/image_classification_mobilenet2_model/main.py
@@ -24,12 +24,16 @@ def load_labels():
 def preprocess_image(image: np.ndarray) -> np.ndarray:
     img = cv2.resize(image, (224, 224))
     img = img.astype(np.float32) / 255.0
-    mean = np.array([0.485, 0.456, 0.406])
-    std = np.array([0.229, 0.224, 0.225])
+
+    # Cast mean and std to float32 to prevent upcasting
+    mean = np.array([0.485, 0.456, 0.406], dtype=np.float32)
+    std = np.array([0.229, 0.224, 0.225], dtype=np.float32)
+
     img = (img - mean) / std
     img = np.transpose(img, (2, 0, 1))  # HWC → CHW
     img = np.expand_dims(img, axis=0)  # Add batch dimension
     return img
+
 
 # Run inference using ONNX
 def predict(image: np.ndarray, session, input_name, output_name, labels) -> str:
